+++ exporter/otlp/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History: opentelemetry-exporter-otlp
 
+### v0.28.1 / 2024-07-24
+
+* ADDED: Improve SSL error logging.
+
+### v0.28.0 / 2024-06-19
+
+* ADDED: Bump google_protobuf >=3.18, < 5.a
+
+### v0.27.0 / 2024-04-19
+
+* ADDED: Add stats for serialization time in otlp exporter
+
 ### v0.26.3 / 2024-02-01
 
 * FIXED: do not log request failure in backoff?